rename #157: DTO 클래스 패키지 구조 변경에 따른 import 문 수정
