+++ src/test/java/fastcampus/team7/Livable_officener/controller/LoginControllerTest.java
@@ -3,8 +3,8 @@ package fastcampus.team7.Livable_officener.controller;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.net.HttpHeaders;
 import fastcampus.team7.Livable_officener.domain.User;
-import fastcampus.team7.Livable_officener.dto.LoginDTO;
-import fastcampus.team7.Livable_officener.dto.LoginDTO.LoginRequestDTO;
+import fastcampus.team7.Livable_officener.dto.login.LoginDTO;
+import fastcampus.team7.Livable_officener.dto.login.LoginDTO.LoginRequestDTO;
 import fastcampus.team7.Livable_officener.global.exception.InvalidPasswordException;
 import fastcampus.team7.Livable_officener.global.exception.NotFoundUserException;
 import fastcampus.team7.Livable_officener.global.sercurity.JwtProvider;
@@ -20,9 +20,9 @@ import org.springframework.http.MediaType;
 import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.web.servlet.MockMvc;
 
-import static fastcampus.team7.Livable_officener.dto.BuildingWithCompaniesDTO.BuildingResponseDTO;
-import static fastcampus.team7.Livable_officener.dto.BuildingWithCompaniesDTO.BuildingWithCompaniesResponseDTO.CompanyResponseDTO;
-import static fastcampus.team7.Livable_officener.dto.LoginDTO.LoginResponseDTO;
+import static fastcampus.team7.Livable_officener.dto.signup.BuildingWithCompaniesDTO.BuildingResponseDTO;
+import static fastcampus.team7.Livable_officener.dto.signup.BuildingWithCompaniesDTO.BuildingWithCompaniesResponseDTO.CompanyResponseDTO;
+import static fastcampus.team7.Livable_officener.dto.login.LoginDTO.LoginResponseDTO;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.BDDMockito.given;
