+++ src/test/java/fastcampus/team7/Livable_officener/service/LoginServiceTest.java
@@ -3,8 +3,8 @@ package fastcampus.team7.Livable_officener.service;
 import fastcampus.team7.Livable_officener.domain.Building;
 import fastcampus.team7.Livable_officener.domain.Company;
 import fastcampus.team7.Livable_officener.domain.User;
-import fastcampus.team7.Livable_officener.dto.LoginDTO;
-import fastcampus.team7.Livable_officener.dto.LoginDTO.LoginRequestDTO;
+import fastcampus.team7.Livable_officener.dto.login.LoginDTO;
+import fastcampus.team7.Livable_officener.dto.login.LoginDTO.LoginRequestDTO;
 import fastcampus.team7.Livable_officener.global.exception.InvalidPasswordException;
 import fastcampus.team7.Livable_officener.global.exception.NotFoundUserException;
 import fastcampus.team7.Livable_officener.global.sercurity.JwtProvider;
